10-at-a-time: typecast the argument passed to sleep()

+++ docs/examples/10-at-a-time.c
@@ -155,7 +155,7 @@ int main(void)
 #ifdef WIN32
         Sleep(L);
 #else
-        sleep(L / 1000);
+        sleep((unsigned int)L / 1000);
 #endif
       }
       else {
